requirements.txt
pyautogen
openai
requests
python-dotenv
tweepy
magnetic-cli 